refactor logic for creating a new game into NewGameButton component
